RDA SymmetricTextEncryption2
RoboCrypto's bigger sister... Implements 3 steps of encryption through plain-text in order to prevent readability. This is not the final product of Project RDA, is it continues to be made plenty more secure.

Algorithm Consists of 3 Steps:
ReverseBlocks();
Rfunction();
MatrixDiffusion();

If you look at the code, it is far from the levels of Advanced Encryption Standard.